Demote ArrayBracketPair to helper class status
