+++ javaparser-testing/src/test/java/com/github/javaparser/bdd/steps/ExistenceOfParentNodeVerifier.java
@@ -498,12 +498,6 @@ class ExistenceOfParentNodeVerifier {
         }
 
         @Override
-        public void visit(ArrayBracketPair n, Void arg) {
-            assertParentIsSet(n);
-            super.visit(n, arg);
-        }
-
-        @Override
         public void visit(NodeList n, Void arg) {
             assertParentIsSet(n);
             super.visit(n, arg);
